Allowing for meltdown overrides for HP updates
A new options --meltdown is added to allow
users to either deploy a pre meltdown or a meltdown
mitigated firmware. By default, a pre meltdown
firmware is installed for the update_hp_firmware.py script.

When using the playbook, the override firmware_use_meltdown
can be enabled to trigger the firmware update using
the lates firmware