Add openstack global requirement constraints
When building the venv for the ops tools, the old style
of limiting contraints to via pip overrides is no longer working.
The pip packages are now clamped by the openstack global requirements
with a hash derived from the OSA service configuration.
If the hash can not be looked up, the versions are limited to Stein.