Update build-ironic-images.sh
Removing devuser in favor of the more secure dynamic-login element

This element allows the operator to inject a SSH key and/or change the root password dynamically when the image boots. Two kernel command line parameters are used to do it:

sshkey
If the operator append sshkey=”$PUBLIC_SSH_KEY” to the kernel command line on boot, the helper script will append this key to the root user authorized_keys.

rootpwd
If the operator append rootpwd=”$ENCRYPTED_PASSWORD” to the kernel command line on boot, the helper script will set the root password to the one specified by this option. Note that this password must be encrypted. Encrypted passwords can be generated using the openssl command, e.g: openssl passwd -1.